% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irwinhall.R
\name{irwinhall}
\alias{irwinhall}
\title{Create Irwin-Hall distribution object}
\usage{
irwinhall(n)
}
\arguments{
\item{n}{number of independent random variables uniformly distributed from 0 to 1, n is a positive integer}
}
\value{
an object of irwinhall class
}
\description{
Create Irwin-Hall distribution object
}
\examples{
ih1 <- irwinhall(5)
ih1$mean
}
