% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirwinhall.R
\name{pirwinhall}
\alias{pirwinhall}
\title{Irwin-Hall cdf}
\usage{
pirwinhall(x, n)
}
\arguments{
\item{x}{random variable, a positive real number <= n}

\item{n}{number of independent random variables uniformly distributed from 0 to 1, n is a positive integer}
}
\value{
numeric value of cumulative distribution function for given x and n
}
\description{
Irwin-Hall cdf
}
\examples{
pirwinhall(2.2, 5)

}
