% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIH.irwinhall.R
\name{plot.irwinhall}
\alias{plot.irwinhall}
\title{Plot Irwin-Hall distribution}
\usage{
\method{plot}{irwinhall}(x, dist = c("pdf", "cdf"), ...)
}
\arguments{
\item{x}{an object of irwinhall class}

\item{dist}{"pdf" (default) for plotting probability density function, "cdf" for cumulative distribution function}

\item{...}{other parameters for plot function}
}
\value{
nothing is returned; plot is produced
}
\description{
Plot Irwin-Hall distribution
}
\examples{
plot(irwinhall(5))
plot(irwinhall(10),dist="cdf")
plot(irwinhall(15),col="red")
}
\seealso{
\code{\link{irwinhall}}, \code{\link{dirwinhall}}, \code{\link{pirwinhall}}
}
