% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIH.irwinhall.R
\name{plotIH.irwinhall}
\alias{plotIH.irwinhall}
\title{Irwin-Hall distribution plot}
\usage{
\method{plotIH}{irwinhall}(obj, dist = c("pdf", "cdf"), ...)
}
\arguments{
\item{obj}{an object of irwinhall class}

\item{dist}{"pdf" for plotting probability density function, "cdf" for cumulative distribution function}

\item{...}{other parameters for plot function}
}
\value{
nothing
}
\description{
Irwin-Hall distribution plot
}
\examples{
plotIH.irwinhall(irwinhall(5),dist="pdf")
}
