% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rirwinhall.R
\name{rirwinhall}
\alias{rirwinhall}
\title{Irwin-Hall randomizer}
\usage{
rirwinhall(num, n)
}
\arguments{
\item{num}{number of observations}

\item{n}{number of independent random variables uniformly distributed from 0 to 1, n is a positive integer}
}
\value{
'num' random values from Irwin-Hall distribution for given n
}
\description{
Irwin-Hall randomizer
}
\examples{
rirwinhall(10, 5)

}
